Restructure core modules and scripts
