+++ Vanta/core/VantaCognitiveEngine.py
