+++ docs/diagrams/component_gridformer.puml
@@ -1,0 +1,7 @@
+@startuml
+package GridFormer {
+  class PulseSmith
+  class GridFormerConnector
+}
+PulseSmith --> GridFormerConnector
+@enduml
