+++ scripts/vmb_advanced_demo.py
