+++ scripts/vmb_advanced_demo_fixed.py
