+++ scripts/vmb_final_status.py
